Widen remaining 32-bit types in send/file-size paths to size_t/FILE_OFFSET
- WebserverSend: int len/bytes_send → size_t
- WebserverSSLSendNonBlocking: same for all 4 SSL backends
- sendData: int ret → size_t
- handleClientWriteDataNotCachedReadWrite: uint32_t diff → FILE_OFFSET
- system_file_access_utils: uint32_t length → FILE_OFFSET for etag,
  enlarge etag buffer from 26 to 34 bytes, uint64_t diff → FILE_OFFSET
- reverse_proxy: int to_send/sent → size_t
- websocket_handler: int send_bytes/to_send → size_t

+++ lib/include/intern/system_sockets.h
@@ -50,7 +50,7 @@ extern "C" {
 
 	int		WebserverAccept(void);
 	int   	WebserverRecv(socket_info* sock,unsigned char *buf, int len, int flags);
-	SOCKET_SEND_STATUS 	WebserverSend(socket_info* sock,const unsigned char *buffer, int len, int flags,int* bytes_send);
+	SOCKET_SEND_STATUS 	WebserverSend(socket_info* sock,const unsigned char *buffer, size_t len, int flags, size_t* bytes_send);
 	int   	WebserverCloseSocket(socket_info* sock);
