+++ lib/include/intern/webserver_ssl.h
@@ -34,7 +34,7 @@ int					WebserverSSLInit(socket_info* s);
 int					WebserverSSLAccept(socket_info* s);
 
 int 				WebserverSSLPending(socket_info* s);
-SOCKET_SEND_STATUS	WebserverSSLSendNonBlocking(socket_info* s, const unsigned char *buf, int len, int flags,int* bytes_send);
+SOCKET_SEND_STATUS	WebserverSSLSendNonBlocking(socket_info* s, const unsigned char *buf, size_t len, int flags, size_t* bytes_send);
 int					WebserverSSLRecvNonBlocking(socket_info* s, unsigned char *buf, unsigned len, int flags);
 int					WebserverSSLCloseSockets(socket_info *s);
